LayoutTests/http/tests/from-origin/redirect-xhr-from-origin-same-blocked-expected.txt

CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/from-origin/resources/xhr.php?fromOrigin=same due to access control checks.
Tests that a redirected cross-origin xhr load fails if the server blocks it with a 'From-Origin: same' response header, even though the server includes an Access-Control-Allow-Origin response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Xhr blocked. 
PASS successfullyParsed is true

TEST COMPLETE
